% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm1D.R
\name{bdm1D}
\alias{bdm1D}
\title{Calculate the Kolmogorov-Chaitin complexity of a string}
\usage{
bdm1D(x, block_size, offset, base)
}
\arguments{
\item{x}{character string}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 1-dimensional Block Decomposition Method (BDM)}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 1-dimensional Block Decomposition Method (BDM)}

\item{base}{number of symbols in the alphabet over which \code{x} is defined}
}
\value{
A number corresponding to the estimated Kolmogorov-Chaitin complexity of the input string.
}
\description{
Calculate the Kolmogorov-Chaitin complexity of a string
}
\details{
This function gives an estimation of the Kolmogorov-Chaitin complexity of a string using the 1-dimensional Block Decomposition Method (BDM).
}
\examples{
\dontrun{
bin_string <- "1001010101000010101001"
bdm1D(bin_string, 12, 11, 2)
}

}
