% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm2D.R
\name{bdm2D}
\alias{bdm2D}
\title{Calculate the Kolmogorov-Chaitin complexity of a matrix}
\usage{
bdm2D(mat, block_size, offset)
}
\arguments{
\item{mat}{a valid matrix}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM)}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM)}
}
\value{
A number corresponding to the estimated Kolmogorov-Chaitin complexity of the input matrix.
}
\description{
Calculate the Kolmogorov-Chaitin complexity of a matrix
}
\details{
This function gives an estimation of the Kolmogorov-Chaitin complexity of a matrix using the 2-dimensional Block Decomposition Method (BDM).
}
\examples{
\dontrun{
set.seed(42)
m99 <- apply(matrix(0, 9, 9), c(1,2), function(x) sample(c(0,1),1))
bdm2D(m99, 3, 3)
}

}
