% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_info_edges.R
\name{calculate_info_edges}
\alias{calculate_info_edges}
\title{Calculate information contribution of edges}
\usage{
calculate_info_edges(x, block_size = 4, offset = 4)
}
\arguments{
\item{x}{an igraph object or a valid adjacency matrix.}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM).}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM).}
}
\value{
A data frame containing the perturbations on the original graph's edges, and their corresponding BDM values and estimated information contributions.
}
\description{
Calculate information contribution of edges
}
\details{
The information contribution of an edge \eqn{e} to a graph \eqn{G} is given by \eqn{I(G,e)=C(G)-C(G-e)}, where \eqn{C(G)} denotes the information content of \eqn{G} and \eqn{C(G-e)} the information content of \eqn{G} after removing \eqn{e}.
This function performs sequential perturbations (deletions) on the edges of \eqn{G} to calculate their information contribution.
}
\examples{
\dontrun{
coxeter_graph <- make_graph("Coxeter")
calculate_info_edges(coxeter_graph)
}

}
