% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_info_symbols.R
\name{calculate_info_symbols}
\alias{calculate_info_symbols}
\title{Calculate information contribution of symbols}
\usage{
calculate_info_symbols(x, block_size, offset, base)
}
\arguments{
\item{x}{character string}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 1-dimensional Block Decomposition Method (BDM).}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 1-dimensional Block Decomposition Method (BDM).}

\item{base}{number of symbols in the alphabet over which \code{x} is defined.}
}
\value{
A data frame containing the perturbations on the original string's characters and their corresponding information contributions.
}
\description{
Calculate information contribution of symbols
}
\details{
The information contribution of a symbol \eqn{i} to a string \eqn{s} is given by \eqn{I(s,i)=C(s)-C(s-i)}, where \eqn{C(s)} denotes the information content of \eqn{s} and \eqn{C(s-i)} the information content of \eqn{s} after removing \eqn{i}.
}
\examples{
\dontrun{
ex_str <- "1010010111010"
calculate_info_symbols(ex_str, 12, 11, 2)
}

}
