% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_info_vertices.R
\name{calculate_info_vertices}
\alias{calculate_info_vertices}
\title{Calculate information contribution of vertices}
\usage{
calculate_info_vertices(x, block_size = 4, offset = 4)
}
\arguments{
\item{x}{an igraph object or a valid adjacency matrix.}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM).}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM).}
}
\value{
A data frame containing the perturbations on the original graph's vertices, and their corresponding BDM values and estimated information contributions.
}
\description{
Calculate information contribution of vertices
}
\details{
The information contribution of a vertex \eqn{v} to a graph \eqn{G} is given by \eqn{I(G,v)=C(G)-C(G-v)}, where \eqn{C(G)} denotes the information content of \eqn{G} and \eqn{C(G-v)} the information content of \eqn{G} after removing \eqn{v}.
This function performs sequential perturbations (deletions) on the vertices of \eqn{G} to calculate their information contribution.
}
\examples{
\dontrun{
tutte_graph <- make_graph("Tutte")
calculate_info_vertices(tutte_graph)
}

}
