% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_mutate.R
\name{element_mutate}
\alias{element_mutate}
\title{Mutate elements of a string}
\usage{
element_mutate(x, index, mutations)
}
\arguments{
\item{x}{character string.}

\item{index}{vector containing the positions of the elements to be replaced.}

\item{mutations}{vector containing the replacements.}
}
\value{
A character string with changes specified by the user at desired positions.
}
\description{
Mutate elements of a string
}
\details{
This function allows one to replace the characters of a given string at certain positions with new characters.
In case the string is binary, the default replacement is just the negation of the bits at each desired locations.
}
\examples{
\dontrun{
# For a binary string
element_mutate("1001", c(1,2,3,4))
# For a non-binary string:
element_mutate("abcd", c(1,2), c("b","a"))
}

}
