% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry_shuffle.R
\name{entry_shuffle}
\alias{entry_shuffle}
\title{Entry shuffle}
\usage{
entry_shuffle(x, row_position, column_position, replacement)
}
\arguments{
\item{x}{a matrix.}

\item{row_position}{an integer indicating the row's index.}

\item{column_position}{an integer indicating the column's index.}

\item{replacement}{the entry to be inserted into the matrix.}
}
\value{
A matrix with a modified entry.
}
\description{
Entry shuffle
}
\details{
Replace a matrix entry.
}
\examples{
\dontrun{
B <- matrix(numeric(15), nrow = 5, ncol = 3, byrow = TRUE)
entry_shuffle(B, 3, 2, -1)
}

}
