% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_signature.R
\name{info_signature}
\alias{info_signature}
\title{Compute the information signature}
\usage{
info_signature(x, what, block_size = 4, offset = 4)
}
\arguments{
\item{x}{an igraph object.}

\item{what}{graph elements to be removed (i.e. edges or vertices).}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM).}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM).}
}
\value{
A data frame corresponding to the information signature of the input graph.
}
\description{
Compute the information signature
}
\details{
The information signature \eqn{\sigma(G)} of graph \eqn{G} is the distribution of information values for either the edges or vertices of \eqn{G}, sorted decreasingly by information contribution.
Information contributions above and below zero respectively correspond to elements whose removal makes the graph lose or gain information.
Neutral elements are, on the other hand, those whose removal has no effect on the graph's information content.
This function is simply a ranking of \code{info_spectra}.
}
\examples{
\dontrun{
# Compute the information signature of the Folkman graph
folkman <- make_graph("Folkman")
is <- info_signature(folkman)
}

}
