% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_spectra.R
\name{info_spectra}
\alias{info_spectra}
\title{Compute the information spectra}
\usage{
info_spectra(x, what, block_size = 4, offset = 4)
}
\arguments{
\item{x}{an igraph object}

\item{what}{graph elements to be removed (i.e. edges or vertices)}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM)}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM)}
}
\value{
A data frame corresponding to the information spectra of the input graph.
}
\description{
Compute the information spectra
}
\details{
The information spectra gives the list of information values for the edges or vertices of a graph.
}
\examples{
\dontrun{
frank <- make_graph("Franklin")
info_spectra(frank, what = 'vertices')
}

}
