% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inforank.R
\name{inforank}
\alias{inforank}
\title{Algorithmic Information Ranking (InfoRank)}
\usage{
inforank(x, block_size, offset, base, what)
}
\arguments{
\item{x}{string, igraph object or valid adjacency matrix.}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM).}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM).}

\item{base}{number of distinct characters in \code{x} when \code{x} is a string.}

\item{what}{graph elements to be removed (i.e. edges or vertices).}
}
\value{
A data frame corresponding to a ranking of the elements of the input object according to their information contribution.
}
\description{
Algorithmic Information Ranking (InfoRank)
}
\details{
This function produces a ranking of the input graph's edges from least informative to most informative edge, i.e. a list of edges sorted in increasing order by their information contribution to the graph.
}
\examples{
\dontrun{
# For strings
inforank("ababbab", base = 2)
# For graphs
g <- make_graph("Smallestcyclicgroup")
inforank(g, what = 'edges')
}

}
