% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info_signature.R
\name{plot_cutting_places}
\alias{plot_cutting_places}
\title{Plot the cutting places}
\usage{
plot_cutting_places(x, what, block_size = 4, offset = 4)
}
\arguments{
\item{x}{an igraph object}

\item{what}{graph elements to be removed (i.e. edges or vertices)}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM)}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM)}
}
\description{
Plot the cutting places
}
\details{
This function plots the differences of consecutive values of the information signature.
}
\examples{
\dontrun{
meredith_graph = make_graph('Meredith')
plot_cutting_places(meredith_graph)
}

}
