% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info_signature.R
\name{plot_info_signature}
\alias{plot_info_signature}
\title{Plot information signature}
\usage{
plot_info_signature(x, what, block_size = 4, offset = 4)
}
\arguments{
\item{x}{an igraph object}

\item{what}{graph elements to be removed (i.e. edges or vertices)}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM)}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM)}
}
\description{
Plot information signature
}
\details{
This function plots the information signature of a given graph.
}
\examples{
\dontrun{
frucht_graph = make_graph('Frucht')
plot_info_signature(frucht_graph, 'edges')
}

}
