% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_shift.R
\name{random_shift}
\alias{random_shift}
\title{Random shift}
\usage{
random_shift(x, position, seed, elements)
}
\arguments{
\item{x}{string or matrix.}

\item{position}{positions of elements to be replaced.}

\item{seed}{seed state.}

\item{elements}{vector corresponding to the alphabet from which the replacement elements are taken from.}
}
\value{
A string or matrix with pseudorandomly generated replacements of symbols or rows.
}
\description{
Random shift
}
\details{
This function replaces one or more bits of a given string or rows of a given matrix with pseudorandom elements from an alphabet.
If the alphabet is not specified in advance, the default binary alphabet, consisting of "0" and "1", will be used.
}
\examples{
\dontrun{
# For a string
x <- "hola amigo"
random_shift(x, c(2,4,6,8), 157)
# For a matrix
A <- matrix(seq(1:100), nrow = 10, ncol = 10, byrow = TRUE)
random_shift(A, c(1, 3), 56)
}

}
