% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_reprogrammability.R
\name{relative_reprogrammability}
\alias{relative_reprogrammability}
\title{Calculate Relative Reprogrammability}
\usage{
relative_reprogrammability(x, what, block_size = 4, offset = 4)
}
\arguments{
\item{x}{an igraph object.}

\item{what}{graph elements to be removed (i.e. edges or vertices).}

\item{block_size}{block size required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM).}

\item{offset}{offset required to estimate the Kolmogorov-Chaitin complexity of \code{x} using the 2-dimensional Block Decomposition Method (BDM).}
}
\value{
A numeric corresponding to the Relative Reprogrammability index of the input graph.
}
\description{
Calculate Relative Reprogrammability
}
\details{
The Relative Reprogrammability index measures the shape of a graph's information signature.
}
\examples{
\dontrun{
robertson_graph <- make_graph("Robertson")
relative_reprogrammability(robertson_graph, 'edges')
}

}
